pyproject.toml
[tool.poetry]
name = "discord-music-bot"
version = "1.0.0"
description = "A simple music bot for discord servers"
authors = ["TrashRoach <rfsnboris@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.1"
pynacl = "^1.5.0"
yt-dlp = "^2023.10.13"

[tool.poetry.group.dev.dependencies]
poetry-dotenv-plugin = "^0.2.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"